Changes to MPI code to help detect out-of-memory errors
Most important change is probably the use of mmap on linux, instead of new[] (in rfhaps.cpp). The reason is that linux allows memory to be overcommitted - That is, malloc never fails, and the system just hopes you requested more than you really need. In our case, we really *do* need all that memory, so if we try and allocate 100gb in rfhaps.cpp, it's initially not backed by physical memory. Then when memset is called it needs to be backed by physical memory, and memory errors happen (and during a call to memset, which is stupid). Using mmap *should* allow us to detect this and fail with a message that we ran out of memory (as opposed to silently crashing).

-Added gridDimX option to mpestrfMpi. This will decrease the amount of memory required at any one time.
-Replaced missing() in slavempestrf with !exists, missing is only for arguments.
-Parts of the logging output of slavempestrf now go straight to file
-Errors now not supressed in mpestrf code
-Changed signature of rfhapsSpecificDesign
-Updated CUDA build, only use sm_30, older architectures are out-dated
-Output logging message if we allocate a contiguous buffer longer than 1gb
-Added checks on return values of some cudaMalloc calls
-Added sharedMmapArray, which calls mmap instead of new[]